requirements.txt
# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# AI / LLM
anthropic>=0.40.0

# Web scraping
tavily-python>=0.5.0

# File parsing
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
markdown>=3.7.0
striprtf>=0.0.26
Pillow>=11.0.0
pytesseract>=0.3.13

# Cloud storage
aioboto3>=13.0.0

# Database
aiosqlite>=0.20.0

# Data validation
pydantic>=2.0.0

# Config
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
